% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_keys.R
\name{authentication_keys}
\alias{authentication_keys}
\title{Authentication Process for Coinbase Advanced Trading}
\usage{
authentication_keys(
  api_key = "",
  secret_key = "",
  method = c("GET", "POST"),
  reqPath,
  body = ""
)
}
\arguments{
\item{api_key}{The api key assigned to the Advanced Trading Wallet}

\item{secret_key}{The secret key assigned to the Advanced Trading Wallet}

\item{method}{Select whether to send a \code{GET} or \code{POST} request}

\item{reqPath}{The request path, can omit \code{"/v3/api/brokerage"}}

\item{body}{The body/payload to pass to the API, usually used with method
\code{"POST"}}
}
\value{
The api_key, the cb_signature, and the timestamp to pass to the api.
}
\description{
In this version of the \code{authentication} function, the api key(s) and the
secret key(s) for a wallet are supplied as arguments to the function.
}
