% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_market_ioc_buy.R
\name{market_market_ioc_buy}
\alias{market_market_ioc_buy}
\title{Function for Market Immediate-or-Cancel Order via Advanced Trade API}
\usage{
market_market_ioc_buy(
  client_order_id = "",
  product_id = "BTC-USD",
  quote_size = 0,
  quote_increment = 2
)
}
\arguments{
\item{client_order_id}{either assign, or the program will assign a random
number}

\item{product_id}{What pair to transact, e.g. \code{BTC-USD}}

\item{quote_size}{Amount of quote currency to spend on order. Required for
\code{BUY} orders.}

\item{quote_increment}{The maximum number of decimal places to which the
\code{quote_size} can be expressed}
}
\value{
A string to pass as payload/body for the Market Immediate-or_Cancel
order
}
\description{
Remember,
Base currency is the first listed (e.g., \code{BTC} in \code{BTC-USD})
Quote currency is the last listed (e.g., \code{USD} in \code{BTC-USD})
}
