% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_market_ioc_sell.R
\name{market_market_ioc_sell}
\alias{market_market_ioc_sell}
\title{Function for Market Immediate-or-Cancel Order via Advanced Trade API}
\usage{
market_market_ioc_sell(
  client_order_id = "",
  product_id = "BTC-USD",
  base_size = 1/21000,
  base_increment = 8
)
}
\arguments{
\item{client_order_id}{either assign, or the program will assign a random number}

\item{product_id}{What pair to transact, e.g. \code{BTC-USD}}

\item{base_size}{Amount of base currency to receive from order. Required for
\code{SELL} orders.}

\item{base_increment}{The maximum number of decimal places to which the
\code{base_size} can be expressed}
}
\value{
A string to pass as payload/body for the Market Immediate-or_Cancel
order
}
\description{
Base currency is the first listed (e.g., \code{BTC} in \code{BTC-USD})
Quote currency is the last listed (e.g., \code{USD} in \code{BTC-USD})
}
\details{
The documentation appears to be wrong, market orders must be specified in
quote currencies
}
