% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_order_ID.R
\name{ran_hex_str}
\alias{ran_hex_str}
\title{Generate Random Hex String of Length n}
\usage{
ran_hex_str(n = 4)
}
\arguments{
\item{n}{the number of consecutive strings to put in a hex string}
}
\description{
This creates the building blocks to generate unique client_order_ID.
}
