% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_limit_gtc.R
\name{limit_limit_gtc}
\alias{limit_limit_gtc}
\title{Function for Limit Good-Til-Cancelled Orders via Advanced Trade API}
\usage{
limit_limit_gtc(
  client_order_id = "",
  product_id = c("BTC-USD", "DOGE-USD", "ETH-USD", "LTC-USD"),
  side = c("BUY", "SELL", "buy", "sell", "B", "b", "S", "s"),
  base_size = 0,
  limit_price = 0
)
}
\arguments{
\item{client_order_id}{either assign, or a random number will be assigned automatically}

\item{product_id}{What pair to transact}

\item{side}{"BUY" or "SELL"}

\item{base_size}{quantity to purchase (in terms of )}

\item{limit_price}{maximum to receive for sell, minimum to pay for buy (USD)}
}
\value{
A string to pass as payload/body for the limit Good-Til-Cancelled
}
\description{
Function for Limit Good-Til-Cancelled Orders via Advanced Trade API
}
